--- a/src/tracking.ts
+++ b/src/tracking.ts
@@ -2,7 +2,6 @@
 
 export type Tracker = (id: number, observableValue: any, currentValue: any) => void;
 var activeTracker: Tracker;
-var trackers: Tracker[] = [];
 var nextObservableId = 0;
 
 export var lastWriteVersion = 0;
@@ -16,21 +15,13 @@ export function takeNextObservableId(): number {
     return ++nextObservableId;
 }
 
-export interface ITrackerExecution {
-    execute<T>(action: () => T): T;
-}
+export function executeWithTracker<T>(action: () => T, tracker?: Tracker): T {
+    var currentTracker = activeTracker;
+    activeTracker = tracker;
 
-export function trackingWith(tracker: Tracker): ITrackerExecution {
-    return {
-        execute: function<T>(action: () => T): T {
-            trackers.push(activeTracker);
-            activeTracker = tracker;
-
-            try {
-                return action();
-            } finally {
-                activeTracker = trackers.pop();
-            }
-        }
-    };
+    try {
+        return action();
+    } finally {
+        activeTracker = currentTracker;
+    }
 }